Update Hold Time
Line 48 onwards is the xml file edited, line 1-46 is the STA.java file edited. the hold time data is seen in the terminal. what's leftover is to make an array to calculate the sum and average the data to then go to the last final screen to show the data